components/worker/requirements.txt
dnspython==1.12.0
docker-py==1.8.0
py2-ipaddress==3.4.1
python-etcd==0.4.3
requests==2.9.1
six==1.10.0
urllib3==1.15.1
websocket-client==0.37.0
wheel==0.24.0
